% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAI_model.R
\name{addpData}
\alias{addpData}
\title{add the cell information into pData slot}
\usage{
addpData(object, pdata, pdata.name = NULL)
}
\arguments{
\item{object}{scAi object}

\item{pdata}{cell information to be added}

\item{pdata.name}{the name of column to be assigned}
}
\description{
add the cell information into pData slot
}
