% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAI_model.R
\name{as_matrix}
\alias{as_matrix}
\title{Convert a sparse matrix to a dense matrix}
\usage{
as_matrix(mat)
}
\arguments{
\item{mat}{A sparse matrix}
}
\description{
Convert a sparse matrix to a dense matrix
}
