% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAI_model.R
\name{bimodLikData}
\alias{bimodLikData}
\title{likelood ratio test}
\usage{
bimodLikData(x, xmin = 0)
}
\arguments{
\item{x}{a vector}

\item{xmin}{threshold for the values in the vector}
}
\description{
likelood ratio test
}
