% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAI_model.R
\name{clust2Mat}
\alias{clust2Mat}
\title{Convert membership into an adjacent matrix}
\usage{
clust2Mat(memb)
}
\arguments{
\item{memb}{Membership vector}
}
\description{
Convert membership into an adjacent matrix
}
