% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAI_model.R
\name{create_scAIobject}
\alias{create_scAIobject}
\title{creat a new scAI object}
\usage{
create_scAIobject(raw.data, do.sparse = T)
}
\arguments{
\item{raw.data}{List of raw data matrices, a paired single-cell transcriptomic and epigenomic data}

\item{do.sparse}{whether use sparse format}
}
\description{
creat a new scAI object
}
