% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_A549}
\alias{data_A549}
\title{single cell multi-omics data of A549 cells.}
\format{A list contains two field data:
\describe{
  \item{data}{Single-cell RNA-seq and ATAC-seq data}
  \item{labels}{The time information of cell groups}
  ...
}}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE117089}
}
\usage{
data_A549
}
\description{
A549 dataset containing paired single-cell RNA-seq and ATAC-seq data as well as the collected time information of each cell
}
\keyword{datasets}
