% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_kidney}
\alias{data_kidney}
\title{single cell multi-omics data of kidney cells.}
\format{A list contains two field data:
\describe{
  \item{data}{Single-cell RNA-seq and ATAC-seq data}
  \item{labels}{The cell type information of cell groups}
  ...
}}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE117089}
}
\usage{
data_kidney
}
\description{
kidney dataset containing paired single-cell RNA-seq and ATAC-seq data as well as the label information of each cell from the original paper
}
\keyword{datasets}
