% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_mESC}
\alias{data_mESC}
\title{single cell multi-omics data of mESC cells.}
\format{A list contains two field data:
\describe{
  \item{data}{Single-cell RNA-seq and single-cell methylation data}
  \item{labels}{The culture condition information of cells}
  ...
}}
\source{
\url{https://github.com/bioFAM/MOFA/blob/master/vignettes/MOFA_example_scMT.Rmd}
}
\usage{
data_mESC
}
\description{
mESC dataset containing paired single-cell RNA-seq and single-cell methylation data as well as the culture condition information of each cell
}
\keyword{datasets}
