% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_simulation8}
\alias{data_simulation8}
\title{A simulation dataset consists of five imbalanced cell clusters with five clusters in scRNA-seq data and three clusters in scATAC-seq data}
\format{A list contains two field data:
\describe{
  \item{data}{Simulated single-cell RNA-seq and ATAC-seq data}
  \item{labels}{The true labels of cell groups}
  ...
}}
\source{
\url{https://genomebiology.biomedcentral.com/articles/10.1186/s13059-020-1932-8}
}
\usage{
data_simulation8
}
\description{
A simulation dataset consists of five imbalanced cell clusters with five clusters in scRNA-seq data and three clusters in scATAC-seq data
}
\keyword{datasets}
