% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAI_model.R
\name{DifferentialLRT}
\alias{DifferentialLRT}
\title{likelood ratio test}
\usage{
DifferentialLRT(x, y, xmin = 0)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}

\item{xmin}{threshold for the values in the vector}
}
\description{
likelood ratio test
}
