% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAI_plotting.R
\name{featureHeatmap}
\alias{featureHeatmap}
\title{generate a heatmap for the expression of differential features across different cell groups}
\usage{
featureHeatmap(object, assay, feature.use, group.by, color.use = NULL,
  use.agg = TRUE, rescaling = TRUE, names.show = TRUE,
  size.names = 8)
}
\arguments{
\item{object}{scAI object}

\item{assay}{define an assay to show, e.g., assay = "RNA"}

\item{feature.use}{a vector of features to show}

\item{group.by}{the name of the variable in pData, defining cell groups. cells are grouped together}

\item{color.use}{colors for the cell clusters}

\item{use.agg}{whether use aggregated data}

\item{rescaling}{whether rescale each feature across all the cells}

\item{names.show}{whether show the feature names}

\item{size.names}{the font size of the feature names}
}
\description{
generate a heatmap for the expression of differential features across different cell groups
}
