% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAI_plotting.R
\name{featureScoreVisualization}
\alias{featureScoreVisualization}
\title{visualize cells on the 2D space with features overlayed}
\usage{
featureScoreVisualization(object, feature.use = NULL, feature.scores,
  method = "VscAI", colormap = "RdPu", color.direction = 1,
  nCol = NULL, xlabel = "VscAI1", ylabel = "VscAI2", show.axes = T,
  cell.size = 0.3, show.legend = T, show.legend.combined = F)
}
\arguments{
\item{object}{scAI object}

\item{feature.use}{a vector of features}

\item{feature.scores}{a matrix containing the feature scores}

\item{method}{dimensional reduction method, e.g., VscAI, tsne, umap}

\item{colormap}{RColorbrewer palette to use}

\item{color.direction}{Sets the order of colours in the scale. If 1, the default, colours are as output by RColorBrewer::brewer.pal(). If -1, the order of colours is reversed.}

\item{nCol}{number of columns of the plot}

\item{xlabel}{label shown on x-axis}

\item{ylabel}{label shown on y-axis}

\item{show.axes}{whether show the axes}

\item{cell.size}{the size of points (cells)}

\item{show.legend}{whether show individual legend}

\item{show.legend.combined}{whether just show one legend}
}
\description{
visualize cells on the 2D space with features overlayed
}
