% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAI_plotting.R
\name{featureVisualization}
\alias{featureVisualization}
\title{visualize cells on the 2D space with gene expression or chromatian accessibility overlayed}
\usage{
featureVisualization(object, assay, feature.use, method = "VscAI",
  nCol = NULL, xlabel = "VscAI1", ylabel = "VscAI2",
  cell.size = 0.3, show.legend = T, show.legend.combined = F,
  show.axes = T)
}
\arguments{
\item{object}{scAI object}

\item{assay}{define an assay to show, e.g., assay = "RNA"}

\item{feature.use}{a vector of features}

\item{method}{dimensional reduction method, e.g., VscAI, tsne, umap}

\item{nCol}{number of columns of the plot}

\item{xlabel}{label shown on x-axis}

\item{ylabel}{label shown on y-axis}

\item{cell.size}{the size of points (cells)}

\item{show.legend}{whether show individual legend}

\item{show.legend.combined}{whether just show one legend}

\item{show.axes}{whether show the axes}
}
\description{
visualize cells on the 2D space with gene expression or chromatian accessibility overlayed
}
