% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAI_model.R
\name{findNearbyLoci}
\alias{findNearbyLoci}
\title{find the nearby loci in a given loci set for a given set of genes}
\usage{
findNearbyLoci(genes, loci, width = 50000, species = "mouse")
}
\arguments{
\item{genes}{a given set of genes}

\item{loci}{the background loci set}

\item{width}{the width from TSS}

\item{species}{mouse or human}
}
\description{
find the nearby loci in a given loci set for a given set of genes
}
