% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAI_model.R
\name{getEmbeddings}
\alias{getEmbeddings}
\title{compute the 2D coordinates of embeded cells, genes, loci and factors using VscAI visualization}
\usage{
getEmbeddings(object, genes.embed = NULL, loci.embed = NULL,
  alpha.embed = 1.9, snn.embed = 5)
}
\arguments{
\item{object}{scAI object}

\item{genes.embed}{A vector of genes to be embedded}

\item{loci.embed}{A vector of loci to be embedded}

\item{alpha.embed}{Parameter controlling the distance between the cells and the factors}

\item{snn.embed}{Number of neighbors}
}
\description{
compute the 2D coordinates of embeded cells, genes, loci and factors using VscAI visualization
}
