% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAI_model.R
\name{identifyClusters}
\alias{identifyClusters}
\title{Identify cell clusters}
\usage{
identifyClusters(object, resolution = 1,
  partition.type = "RBConfigurationVertexPartition", seed.use = 42L,
  n.iter = 10L, initial.membership = NULL, weights = NULL,
  node.sizes = NULL, K = NULL)
}
\arguments{
\item{object}{scAI object}

\item{resolution}{A parameter controlling the coarseness of the clusters}

\item{partition.type}{Type of partition to use. Defaults to RBConfigurationVertexPartition. Options include: ModularityVertexPartition, RBERVertexPartition, CPMVertexPartition, MutableVertexPartition, SignificanceVertexPartition, SurpriseVertexPartition (see the Leiden python module documentation for more details)}

\item{seed.use}{set seed}

\item{n.iter}{number of iteration}

\item{initial.membership}{arameters to pass to the Python leidenalg function defaults initial_membership=None}

\item{weights}{defaults weights=None}

\item{node.sizes}{Parameters to pass to the Python leidenalg function}

\item{K}{Number of clusters if performing hierarchical clustering of the consensus matrix}
}
\description{
Identify cell clusters
}
