% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAI_model.R
\name{inferRegulations}
\alias{inferRegulations}
\title{Infer regulatory relationships}
\usage{
inferRegulations(object, gene.use, candinate_loci, cutoff_H = 0.5,
  cutoff = 0.1, thresh_corr = 0.2)
}
\arguments{
\item{object}{scAI object}

\item{gene.use}{genes to be inferred}

\item{candinate_loci}{cadinate loci}

\item{cutoff_H}{threshold of coefficient values}

\item{cutoff}{the difference of correlation to be considered as significant}

\item{thresh_corr}{threshold of correlation coefficient}
}
\description{
Infer regulatory relationships
}
