% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pairedData_simulation}
\alias{pairedData_simulation}
\title{simulated single cell multi-omics data.}
\format{A list contains two field data:
\describe{
  \item{data}{Simulated single-cell RNA-seq and ATAC-seq data}
  \item{labels}{The true labels of cell groups}
  ...
}}
\source{
\url{https://bitbucket.org/ConesaLab/mosim/src/master/}
}
\usage{
pairedData_simulation
}
\description{
A dataset containing paired single-cell RNA-seq and ATAC-seq data as well as the true labels of cell groups
}
\keyword{datasets}
