% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAI_model.R
\name{preprocessing}
\alias{preprocessing}
\title{preprocess the raw.data including quality control and normalization}
\usage{
preprocessing(object, assay = list("RNA", "ATAC"), minFeatures = 200,
  minCells = 3, minCounts = NULL, maxCounts = NULL,
  libararyflag = TRUE, logNormalize = TRUE)
}
\arguments{
\item{object}{scAI object}

\item{assay}{List of assay names to be normalized}

\item{minFeatures}{Filter out cells with expressed features < minimum features}

\item{minCells}{Filter out features expressing in less than minCells}

\item{minCounts}{Filter out cells with expressed count < minCounts}

\item{maxCounts}{Filter out cells with expressed count > minCounts}

\item{libararyflag}{Whether do library size normalization}

\item{logNormalize}{whether do log transformation}
}
\description{
preprocess the raw.data including quality control and normalization
}
