% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAI_model.R
\name{reorderFeatures}
\alias{reorderFeatures}
\title{Reorder features according to the loading values}
\usage{
reorderFeatures(W, cutoff = 0.5)
}
\arguments{
\item{W}{Basis matrix}

\item{cutoff}{Threshold of the loading values}
}
\description{
Reorder features according to the loading values
}
