% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAI_model.R
\name{run_scAI}
\alias{run_scAI}
\title{run scAI model}
\usage{
run_scAI(
  object,
  K,
  do.fast = FALSE,
  nrun = 5L,
  hvg.use1 = FALSE,
  hvg.use2 = FALSE,
  keep_all = F,
  s = 0.25,
  alpha = 1,
  lambda = 1e+05,
  gamma = 1,
  maxIter = 500L,
  stop_rule = 1L,
  init = NULL,
  rand.seed = 1L
)
}
\arguments{
\item{object}{scAI object}

\item{K}{An integer indicating the Rank of the inferred factor}

\item{do.fast}{whether use the python version for running scAI optimization model}

\item{nrun}{Number of times to repreat the running}

\item{hvg.use1}{whether use high variable genes for RNA-seq data}

\item{hvg.use2}{whether use high variable genes for ATAC-seq data}

\item{keep_all}{Whether keep all the results from multiple runs}

\item{s}{Probability of Bernoulli distribution}

\item{alpha}{model parameter}

\item{lambda}{model parameter}

\item{gamma}{model parameter}

\item{maxIter}{An integer indicating Maximum number of iteration}

\item{stop_rule}{Stop rule to be used}

\item{init}{List of the initialized low-rank matrices}

\item{rand.seed}{An integer indicating the seed}
}
\value{

}
\description{
run scAI model
}
