% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAI_model.R
\docType{class}
\name{scAI-class}
\alias{scAI-class}
\alias{scAI}
\title{The scAI Class}
\description{
The scAI object is created from a paired single-cell transcriptomic and epigenomic data.
It takes a list of two digital data matrices as input. Genes/loci should be in rows and cells in columns. rownames and colnames should be included.
The class provides functions for data preprocessing, integrative analysis, and visualization.
}
\details{
The key slots used in the scAI object are described below.
}
\section{Slots}{

\describe{
\item{\code{raw.data}}{List of raw data matrices, one per dataset (Genes/loci should be in rows and cells in columns)}

\item{\code{norm.data}}{List of normalized matrices (genes/loci by cells)}

\item{\code{agg.data}}{Aggregated epigenomic data within similar cells}

\item{\code{scale.data}}{List of scaled matrices}

\item{\code{pData}}{data frame storing the information associated with each cell}

\item{\code{var.features}}{List of informative features to be used, one giving informative genes and the other giving informative loci}

\item{\code{fit}}{List of inferred low-rank matrices, including W1, W2, H, Z, R}

\item{\code{fit.variedK}}{List of inferred low-rank matrices when varying the rank K}

\item{\code{embed}}{List of the reduced 2D coordinates, one per method, e.g., t-SNE/FIt-SNE/umap}

\item{\code{identity}}{a factor defining the cell identity}

\item{\code{cluster}}{List of consensus clustering results}

\item{\code{options}}{List of parameters used throughout analysis}
}}

