% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAI_model.R
\name{scAImodel}
\alias{scAImodel}
\title{Solving the optimization problem in scAI}
\usage{
scAImodel(X1, X2, K, s = 0.25, alpha = 1, lambda = 1e+05,
  gamma = 1, maxIter = 500, stop_rule = 1, R.init = NULL,
  W1.init = NULL, W2.init = NULL, H.init = NULL, Z.init = NULL)
}
\arguments{
\item{X1}{Single-cell transcriptomic data matrix (norm.data)}

\item{X2}{Single-cell epigenomic data matrix (norm.data)}

\item{K}{Rank of inferred factors}

\item{s}{Probability of Bernoulli distribution}

\item{alpha}{model parameter}

\item{lambda}{model parameter}

\item{gamma}{model parameter}

\item{maxIter}{Maximum number of iteration}

\item{stop_rule}{Stop rule to be used}

\item{R.init}{initialization of R}

\item{W1.init}{initialization of W1}

\item{W2.init}{initialization of W2}

\item{H.init}{initialization of H}

\item{Z.init}{initialization of Z}
}
\description{
Solving the optimization problem in scAI
}
