% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAI_model.R
\name{searchGeneRegions}
\alias{searchGeneRegions}
\title{find chromsome regions of genes}
\usage{
searchGeneRegions(genes, species = "mouse")
}
\arguments{
\item{genes}{a given set of genes}

\item{species}{mouse or human}
}
\description{
find chromsome regions of genes
}
