% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAI_model.R
\name{selectFeatures}
\alias{selectFeatures}
\title{select highly variable features}
\usage{
selectFeatures(object, assay = "RNA", do.plot = TRUE, do.text = TRUE,
  x.low.cutoff = 0.01, x.high.cutoff = 3.5, y.cutoff = 0.5,
  y.high.cutoff = Inf, num.bin = 20, pch.use = 16,
  col.use = "black", cex.text.use = 0.5)
}
\arguments{
\item{object}{scAI objecy}

\item{assay}{Name of assay}

\item{do.plot}{Whether showing plot}

\item{do.text}{Whether adding feature names}

\item{x.low.cutoff}{The minimum expression level}

\item{x.high.cutoff}{The maximum expression level}

\item{y.cutoff}{The minimum fano factor values}

\item{y.high.cutoff}{The maximum fano factor values}

\item{num.bin}{Number of bins}

\item{pch.use}{Shape of dots in ggplot}

\item{col.use}{Color of dots}

\item{cex.text.use}{Size of text}
}
\description{
select highly variable features
}
