% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAI_model.R
\name{selectK}
\alias{selectK}
\title{Select number of the rank}
\usage{
selectK(object, rangeK = c(2:15))
}
\arguments{
\item{object}{scAI object}

\item{rangeK}{A predefined range of K}
}
\description{
Select number of the rank
}
