% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAI_model.R
\docType{methods}
\name{show,scAI-method}
\alias{show,scAI-method}
\title{show method for scAI}
\usage{
\S4method{show}{scAI}(object)
}
\arguments{
\item{object}{object}

\item{scAI}{object}

\item{show}{show the object}
}
\description{
show method for scAI
}
