% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAI_plotting.R
\name{VscAIplot}
\alias{VscAIplot}
\title{VscAI visualize the genes, loci and factors that separate cell states on two dimensions alongside the cells}
\usage{
VscAIplot(object, gene.use, loci.use, loci.use.names, color.by,
  labels.order = NULL, colors.use = NULL, brewer.use = FALSE,
  xlabel = "VscAI1", ylabel = "VscAI2", title = NULL,
  label.size = 3, cell.size = 0.3, font.size = 10, do.label = T,
  show.legend = T, show.axes = T)
}
\arguments{
\item{object}{scAI object}

\item{gene.use}{embedded genes}

\item{loci.use}{embedded loci}

\item{loci.use.names}{alternative names of embedded loci, e.g, the corresponding motif}

\item{color.by}{the name of the variable in pData, defining cell groups, cells are colored based on the labels}

\item{labels.order}{defining the factor level}

\item{colors.use}{defining the color for each cell group}

\item{brewer.use}{use RColorBrewer palette instead of default ggplot2 color}

\item{xlabel}{label of x-axis}

\item{ylabel}{label of y-axis}

\item{title}{main title of the plot}

\item{label.size}{font size of the legend}

\item{cell.size}{size of the dots}

\item{font.size}{size of font}

\item{do.label}{label the cluster in 2D space}

\item{show.legend}{whether show the legend}

\item{show.axes}{whether show the axes}
}
\value{
ggplot2 object with 2D plot
}
\description{
VscAI visualize the genes, loci and factors that separate cell states on two dimensions alongside the cells
}
