% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAI_plotting.R
\name{featureHeatmap}
\alias{featureHeatmap}
\title{generate a heatmap for the expression of differential features across different cell groups}
\usage{
featureHeatmap(object, assay, feature.use, group.by, rescaling = T)
}
\arguments{
\item{object}{scAI object}

\item{assay}{define an assay to show, e.g., assay = "RNA"}

\item{feature.use}{a vector of features to show}

\item{group.by}{the name of the variable in pData, defining cell groups. cells are grouped together}

\item{rescaling}{whether rescale each feature across all the cells}
}
\description{
generate a heatmap for the expression of differential features across different cell groups
}
