% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAI_model.R
\name{identifyClusterMarkers}
\alias{identifyClusterMarkers}
\title{Identify enriched features in each cell cluster}
\usage{
identifyClusterMarkers(object, assay, features = NULL,
  test.use = "bimod", thresh.pc = 0.05, thresh.fc = 0.25,
  thresh.p = 0.05, n.top = 10)
}
\arguments{
\item{object}{scAI object}

\item{assay}{Name of assay to be analyzed}

\item{features}{a vector of features}

\item{test.use}{which test to use ("bimod" or "wilcox")}

\item{thresh.pc}{Threshold of the percent of cells enriched in one cluster}

\item{thresh.fc}{Threshold of Fold Change}

\item{thresh.p}{Threshold of p-values}

\item{n.top}{Number of top features to be returned}
}
\description{
Identify enriched features in each cell cluster
}
