% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAI_model.R
\name{run_scAI}
\alias{run_scAI}
\title{run scAI model}
\usage{
run_scAI(object, K, nrun = 5, keep_all = F, s = 0.25, alpha = 1,
  lambda = 1e+05, gamma = 1, maxIter = 500, stop_rule = 1,
  init = NULL, rand.seed = 1)
}
\arguments{
\item{object}{scAI object}

\item{K}{Rank of the inferred factor}

\item{nrun}{Number of times to repreat the running}

\item{keep_all}{Whether keep all the results from multiple runs}

\item{s}{Probability of Bernoulli distribution}

\item{alpha}{model parameter}

\item{lambda}{model parameter}

\item{gamma}{model parameter}

\item{maxIter}{Maximum number of iteration}

\item{stop_rule}{Stop rule to be used}

\item{init}{List of the initialized low-rank matrices}

\item{rand.seed}{seed}
}
\description{
run scAI model
}
